requirements.txt
hydra-core
omegaconf
h5py
numpy
numcodecs
opencv-python
pandas
scipy
zarr
tqdm
einops
wandb
pynput
click
threadpoolctl
matplotlib
dataclasses
atomics
dill
swanlab
pyrealsense2
imagecodecs
pygame
pymunk
pydantic
numba==0.56.4